Begin writing script to extract Psalter from BCP
Getting the Psalter into a format where it can be used by the rest of
program is going to take some work:

1. Get plain text of the Psalter from a PDF of the BCP 2019.
2. Clean up the text and remove formatting.
3. Split the Psalter into individual psalms and verses.
4. Convert everything into JSON for ease of use later.

So far, I've completed step 1 and have started step 2. I'm using the
PyMuPDF library to extract the text from the PDF. Seems like a very
powerful library and worth exploring more.