Start putting everything together
There is still a lot of work to be done (and even more testing), but I
think I'm getting close. I need to write a function that calculates the
next Sunday in the liturgical calendar based on the current date, and
then I need to start writing a nice CLI tool with some helpful options.
After that I'll be pretty much finished with this first draft.