Write methods to get lessons for any day in the liturgical year
This is not a very elegant solution, but I honestly can't think of how
else to go about it. The Gregorian calendar and the Liturgical calendar
do not neatly overlap. Beyond that, the Liturgical calendar is made up
of two major cycles which interact with each other. Unfortunately, the
Christmas cycle revolves around absolute dates in the solar calendar,
while the Easter cycle revolves around moveable dates in the lunar
calendar. In other words, it's a giant pain to get everything working
together.

There are still some edge cases that I haven't covered, but this is a
good start. I'm looking forward to investigating other implementations
when I'm finished with this project. This is definitely one I want to
come back to in the future.

Now, it's finally time to write some tests.