Finish first draft of the program
Everything seems to be working as intended. I think it's just about
ready for real users, but first I want to add a bunch of additional
unit tests, especially for the Lectionary class. After that, all I need
to do is write up the README and I'm ready to call it version 1.0.0 and
release it.